+++ src/app/utils/util.ts
@@ -44,3 +44,33 @@ export function downloadObjectAsJson(exportObj, exportName) {
     downloadAnchorNode.click();
     downloadAnchorNode.remove();
 }
+
+export function secondsToDhms(seconds) {
+    seconds = Number(seconds);
+
+    var d = Math.floor(seconds / (3600*24));
+    var h = Math.floor(seconds / 3600);
+    var m = Math.floor(seconds % 3600 / 60);
+    var s = Math.floor(seconds % 3600 % 60);
+
+    var dDisplay = d > 0 ? d + (d == 1 ? " day, " : " days, ") : "";
+    var hDisplay = h > 0 ? h + (h == 1 ? " hour, " : " hours, ") : "";
+    var mDisplay = m > 0 ? m + (m == 1 ? " minute, " : " minutes, ") : "";
+    var sDisplay = s > 0 ? s + (s == 1 ? " second" : " seconds") : "";
+    let label = dDisplay + hDisplay + mDisplay + sDisplay;
+
+    const getDays= function(): number{
+        return d;
+    }
+    const getHours= function(): number{
+        return h;
+    }
+    const getMinutes= function(): number{
+        return m;
+    }
+    const getSeconds= function(): number{
+        return s;
+    }
+
+    return { getDays: getDays, getHours: getHours,  getMinutes: getMinutes, getSeconds: getSeconds}
+}